+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Ignore updates that arrive after the snapshot, but are timestamped before the snapshot
 
 ## [0.0.6] - 2023-09-23
 ### Fixed